Reviewer comments
I put the code into Titanium and did not see any errors in it. The phone gadget that was used was the map gadget that was found in chapter 7 on page 173. One suggestion that I have is to create your own theme. You used the color codes that were given in the book and I think it would have been better if you chose a theme yourself and came up with the colors to fit into that theme.